Update packet files for packet ID changes
