+++ src/main/java/emu/grasscutter/server/game/GameServerPacketHandler.java
@@ -1,18 +1,16 @@
 package emu.grasscutter.server.game;
 
-import static emu.grasscutter.config.Configuration.GAME_INFO;
-
 import emu.grasscutter.Grasscutter;
 import emu.grasscutter.Grasscutter.ServerDebugMode;
-import emu.grasscutter.net.packet.Opcodes;
-import emu.grasscutter.net.packet.PacketHandler;
-import emu.grasscutter.net.packet.PacketOpcodes;
+import emu.grasscutter.net.packet.*;
 import emu.grasscutter.server.event.game.ReceivePacketEvent;
 import emu.grasscutter.server.game.GameSession.SessionState;
-import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
-import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
-import java.util.Set;
+import it.unimi.dsi.fastutil.ints.*;
 import org.reflections.Reflections;
+
+import java.util.Set;
+
+import static emu.grasscutter.config.Configuration.GAME_INFO;
 
 @SuppressWarnings("unchecked")
 public class GameServerPacketHandler {
