+++ src/main/java/emu/grasscutter/server/game/GameSession.java
@@ -1,8 +1,5 @@
 package emu.grasscutter.server.game;
 
-import static emu.grasscutter.config.Configuration.*;
-import static emu.grasscutter.utils.lang.Language.translate;
-
 import emu.grasscutter.Grasscutter;
 import emu.grasscutter.Grasscutter.ServerDebugMode;
 import emu.grasscutter.game.Account;
@@ -11,10 +8,14 @@ import emu.grasscutter.net.packet.*;
 import emu.grasscutter.server.event.game.SendPacketEvent;
 import emu.grasscutter.utils.*;
 import io.netty.buffer.*;
+import lombok.*;
+
 import java.io.File;
 import java.net.InetSocketAddress;
 import java.nio.file.Path;
-import lombok.*;
+
+import static emu.grasscutter.config.Configuration.*;
+import static emu.grasscutter.utils.lang.Language.translate;
 
 public class GameSession implements GameSessionManager.KcpChannel {
     private final GameServer server;
