+++ src/main/java/emu/grasscutter/server/packet/send/PacketFireworksLaunchDataNotify.java
@@ -1,25 +1,13 @@
-// removed due to proto not existing
-
-/*
 package emu.grasscutter.server.packet.send;
 
-import emu.grasscutter.net.packet.BasePacket;
-import emu.grasscutter.net.packet.PacketOpcodes;
-import emu.grasscutter.net.proto.FireworksLaunchDataNotifyOuterClass;
-import emu.grasscutter.net.proto.FireworksLaunchSchemeDataOuterClass;
+import emu.grasscutter.net.packet.*;
+import emu.grasscutter.net.proto.FireworksLaunchDataNotifyOuterClass.FireworksLaunchDataNotify;
+import emu.grasscutter.net.proto.FireworksLaunchSchemeDataOuterClass.FireworksLaunchSchemeData;
 
 public class PacketFireworksLaunchDataNotify extends BasePacket {
-
-    public PacketFireworksLaunchDataNotify(FireworksLaunchSchemeDataOuterClass.FireworksLaunchSchemeData notify) {
+    public PacketFireworksLaunchDataNotify(FireworksLaunchSchemeData data) {
         super(PacketOpcodes.FireworksLaunchDataNotify);
 
-        var proto
-                = FireworksLaunchDataNotifyOuterClass.FireworksLaunchDataNotify.newBuilder();
-
-        proto.setLastUseSchemeId(1).addSchemeDataList(notify);
-
-        setData(proto.build());
+        this.setData(FireworksLaunchDataNotify.newBuilder().addSchemeDataList(data));
     }
-
 }
-*/
