+++ src/main/java/emu/grasscutter/server/packet/send/PacketFireworksReformDataNotify.java
@@ -1,25 +1,14 @@
-// removed due to proto not existing
-
-/*
 package emu.grasscutter.server.packet.send;
 
-import emu.grasscutter.net.packet.BasePacket;
-import emu.grasscutter.net.packet.PacketOpcodes;
-import emu.grasscutter.net.proto.FireworksReformDataNotifyOuterClass;
-import emu.grasscutter.net.proto.FireworksReformDataOuterClass;
+import emu.grasscutter.net.packet.*;
+import emu.grasscutter.net.proto.FireworksReformDataNotifyOuterClass.FireworksReformDataNotify;
+import emu.grasscutter.net.proto.FireworksReformDataOuterClass.FireworksReformData;
 
 public class PacketFireworksReformDataNotify extends BasePacket {
 
-    public PacketFireworksReformDataNotify(FireworksReformDataOuterClass.FireworksReformData fireworksReformData) {
+    public PacketFireworksReformDataNotify(FireworksReformData data) {
         super(PacketOpcodes.FireworksReformDataNotify);
 
-        var proto
-            = FireworksReformDataNotifyOuterClass.FireworksReformDataNotify.newBuilder();
-
-        proto.addFireworksReformDataList(fireworksReformData);
-
-        setData(proto.build());
+        this.setData(FireworksReformDataNotify.newBuilder().addFireworksReformDataList(data));
     }
-
 }
-*/
